Implementación de vacunación, inmunidad y otras mejoras
Actualizado archivo con variación de beta 2021.

- InfectionReport
Diferenciados nuevos casos por grupo etario: sintomáticos y asintomáticos.
Gráfica de nuevos casos diarios, cambiada para solo casos sintomáticos.
Agregados contadores de primera y segunda dosis aplicadas.
Nueva gráfica de primera y segundas dosis aplicadas acomuladas.

- BuildingManager
Limpieza, eliminados métodos sin uso.

- DataSet
"SPACE_BETWEEN_STUDENTS" incrementada distancia entre pupitres en escuelas, para disminuir el contagio.
"DEFAULT_OOC_CONTAGION_VALUE" ajustado por cambios en cálculo.
"INFECTION_RATE" movido desde clase "Temperature".
"CS_MIN_INFECTION_RATE" ajustado por cambios en fórmula.
"DEFAULT_PREICU_DEATH_RATE" nueva chance de muerte pre-UTI.
"DEFAULT_ICU_DEATH_RATE" ajustado a 55%.
Nuevos atributos para la implementación de niveles, periodos de inmunidad y vacunación.

- PlaceProperty
Actualizado uso de métodos de lectura de archivos.

- Temperature
Implementada diferenciación entre años bisiestos, se toman 365 o 366 valores según año.
Actualizada fórmula para calcular beta fuera de contexto (OOC).
Actualizado uso de métodos de lectura de archivos.

- Town
Nuevos métodos para obtener ruta de archivos de vacunación de cada ciudad.

- BuildingAgent
Ahora se eliminan las instancias de SurfaceAgent cuando dejan de estar contaminadas.
Otras optimizaciones menores.

- ClassroomAgent
Metodo "lookForCloseContacts" ahora se fija que "contacto estrecho" este habilitado,
en lugar del protocolo burbuja. Esto es para que se cumpla incluso con "presencialidad plena".

- HumanAgent
Agregado estado "deceased" (fallecido).
Nuevos atributos y métodos para la implementación de niveles y duración de inmunidad (natural o por vacuna).
Nuevo método "checkContagion" para confirmar contagio, reemplaza a chequeos previos de contagio.
Agregada implementación de fallecimiento pre-UTI. Para diferenciar se cambia el término por caso grave.
Algunas optimizaciones menores.

- SurfaceAgent
Cálculo de grado de contaminación (o beta) simplificado.
Algunas optimizaciones menores.

- SubContext
Nuevos métodos para la manipulación de agentes humanos en sub contexto.
Instancia e inicializa "VaccinationManager" para realizar la vacunación local.
Fix en metodo "removeSchoolProtocolAction" que deshabilitaba protocolo burbuja,
antes de eliminar correctamente la acción programada "startSchoolProtocolAction".
Algunos atributos renombrados.
Limpieza, eliminados métodos sin uso.

- VaccinationManager
Nueva clase encargada de gestionar el calendario de vacunación.
Implementa métodos de aplicación de dosis y cálculo de inmunidad, entre otras funciones.

- Utils
Nueva clase para unificar la implementación de métodos de lectura de archivos CSV y otros.

+++ src/geocovid/agents/WorkplaceAgent.java
@@ -25,7 +25,7 @@ public class WorkplaceAgent extends BuildingAgent {
 	/** Lista de HumanAgent pre-trasmisores (sintomaticos) */
 	protected List<HumanAgent> preSpreadersList = new ArrayList<HumanAgent>();
 	
-	public WorkplaceAgent(SubContext subContext, int sectoralType, int sectoralIndex, Coordinate coord, long id, String workType, int activityType, int area, int coveredArea, int workersPlace, int workersArea) {
+	public WorkplaceAgent(SubContext subContext, int sectoralType, int sectoralIndex, Coordinate coord, int id, String workType, int activityType, int area, int coveredArea, int workersPlace, int workersArea) {
 		super(subContext, sectoralType, sectoralIndex, coord, id, workType, area, (area * coveredArea) / 100, DataSet.WORKPLACE_AVAILABLE_AREA, true);
 		
 		this.workplaceType = workType;
@@ -39,7 +39,7 @@ public class WorkplaceAgent extends BuildingAgent {
 		createWorkPositions();
 	}
 
-	public WorkplaceAgent(SubContext subContext, int sectoralType, int sectoralIndex, Coordinate coord, long id, String workType, int activityType, int width, int lenght, int workersPlace) {
+	public WorkplaceAgent(SubContext subContext, int sectoralType, int sectoralIndex, Coordinate coord, int id, String workType, int activityType, int width, int lenght, int workersPlace) {
 		super(subContext, sectoralType, sectoralIndex, coord, id, workType, width, lenght, true);
 		
 		this.workplaceType = workType;
@@ -219,7 +219,7 @@ public class WorkplaceAgent extends BuildingAgent {
 			}
 		}
 		// Chequea efectividad de cubrebocas
-		if (context.getMaskEffectivity() > 0) {
+		if (context.getMaskEffectivity() > 0d) {
 			// Si es un lugar cerrado o se usa cubreboca en exteriores
 			if (!isOutdoor() || context.wearMaskOutdoor()) {
 				// Si el contagio es entre cliente-cliente/empleado-cliente o entre empleados
@@ -227,11 +227,11 @@ public class WorkplaceAgent extends BuildingAgent {
 				if (workers) {
 					// Chequea si utilizan cubreboca entre empleados del tipo de actividad de parcela
 					if ((activityType == 1 && context.wearMaskWorkspace()) || (activityType > 1 && context.wearMaskCS()))
-						infectionRate *= 1 - context.getMaskEffectivity();
+						infectionRate *= 1d - context.getMaskEffectivity();
 				}
 				// Chequea si utilizan cubreboca entre clientes y empleados
 				else if (context.wearMaskCustomer()) {
-					infectionRate *= 1 - context.getMaskEffectivity();
+					infectionRate *= 1d - context.getMaskEffectivity();
 				}
 			}
 		}
